pyproject.toml
[build-system]
requires = ["hatchling>=1.21.0"]
build-backend = "hatchling.build"

[project]
name = "meridian"
version = "0.1.0"
description = "MERIDIAN Pathfinder: gated ML decision engineering framework"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "Proprietary" }
authors = [{ name = "Vincent Gibson" }]
dependencies = [
  "typer>=0.12.0",
  "pydantic>=2.0.0",
  "pydantic-settings>=2.0.0",
  "pyyaml>=6.0.0",
  "rich>=13.0.0",
  "pandas>=2.0.0",
  "numpy>=1.24.0",
  "scikit-learn>=1.3.0",
  "anthropic>=0.40.0",
  "httpx>=0.27.0",
  "pytest>=8.0.0",
]

[project.scripts]
meridian = "meridian.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["meridian"]

